Use decimal cause it can store both decimal and integers, and force stat values type.

+++ SteamRetriever/Models/AchievementModels.cs
@@ -88,5 +88,5 @@ public class StatModel
     public object Default { get; set; }
 
     [JsonProperty("windowsize")]
-    public double? WindowSize { get; set; }
+    public decimal? WindowSize { get; set; }
 }
